Add files via upload
Graph of the temporal variability of the daily Google-derived mobility categories, the daily COVID-19 cases & death rates from OWID for Ireland (n=333) in relation to government announcements, including travel 
restrictions and lockdowns/stay-at-home orders, national holidays and meteorological conditions from 17/02/2020 to 15/01/2021. The data used is the raw unprocessed Ireland data. The Google-derived mobility categories are 
Retail & Recreation, Grocery & Pharmacy Stores, Parks, Transit Stations, Workplaces and Residential, and they are represented as the % change in relation to a 5 week baseline ranging from 03/01/2020 to 06/02/2020. The 3 
waves of COVID-19 in Ireland are identifiable by the colour blocks, which end at the completion of a lockdown period. Wave 1 ranges between 17/02/2020 and 28/06/2020, wave 2 ranges between 28/06/2020 and 01/12/2020 
and wave 3 ranges between 01/12/2020 and 15/01/2021